backend/src/services/auth_service.py
from datetime import datetime, timedelta
from typing import Optional
import uuid
from sqlmodel import Session
from jose import JWTError, jwt
from ..models.user import User, UserCreate
from ..config import settings
from ..utils.password_utils import verify_password
from ..utils.db_utils import get_user_by_email

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """Create a JWT access token"""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)

    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.secret_key, algorithm=settings.algorithm)
    return encoded_jwt

def authenticate_user(session: Session, email: str, password: str) -> Optional[User]:
    """Authenticate a user by email and password"""
    user = get_user_by_email(session, email)
    if not user or not verify_password(password, user.password_hash):
        return None
    return user

